Fixed some tests in hs_core to use valid countries.
Also, PEP8 of the relevant files.

+++ hs_core/views/resource_rest_api.py
@@ -23,8 +23,7 @@ from rest_framework import status
 from hs_core import hydroshare
 from hs_core.models import AbstractResource
 from hs_core.hydroshare.utils import get_resource_by_shortkey, get_resource_types
-from hs_core.views import utils as view_utils
-from hs_core.views.utils import ACTION_TO_AUTHORIZE
+from hs_core.views.utils import authorize, ACTION_TO_AUTHORIZE
 from hs_core.views import serializers
 from hs_core.views import pagination
 from hs_core.hydroshare.utils import get_file_storage, resource_modified
@@ -256,8 +255,8 @@ class ResourceReadUpdateDelete(ResourceToListItemMixin, generics.RetrieveUpdateD
     def get(self, request, pk):
         """ Get resource in zipped BagIt format
         """
-        res, _, _ = view_utils.authorize(request, pk,
-                                         needed_permission=ACTION_TO_AUTHORIZE.VIEW_RESOURCE)
+        res, _, _ = authorize(request, pk,
+                              needed_permission=ACTION_TO_AUTHORIZE.VIEW_RESOURCE)
         site_url = hydroshare.utils.current_site_url()
         if res.resource_type.lower() == "reftimeseriesresource":
 
@@ -275,7 +274,7 @@ class ResourceReadUpdateDelete(ResourceToListItemMixin, generics.RetrieveUpdateD
 
     def delete(self, request, pk):
         # only resource owners are allowed to delete
-        view_utils.authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.DELETE_RESOURCE)
+        authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.DELETE_RESOURCE)
         hydroshare.delete_resource(pk)
         # spec says we need return the id of the resource that got deleted - otherwise would
         # have used status code 204 and not 200
@@ -455,7 +454,7 @@ class SystemMetadataRetrieve(ResourceToListItemMixin, APIView):
     def get(self, request, pk):
         """ Get resource system metadata, as well as URLs to the bag and science metadata
         """
-        view_utils.authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.VIEW_METADATA)
+        authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.VIEW_METADATA)
         res = get_resource_by_shortkey(pk)
         ser = self.get_serializer_class()(self.resourceToResourceListItem(res))
 
@@ -482,7 +481,7 @@ class AccessRulesUpdate(APIView):
         """ Update access rules
         """
         # only resource owners are allowed to change resource flags (e.g., public)
-        view_utils.authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.SET_RESOURCE_FLAG)
+        authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.SET_RESOURCE_FLAG)
 
         access_rules_validator = serializers.AccessRulesRequestValidator(data=request.data)
         if not access_rules_validator.is_valid():
@@ -533,14 +532,14 @@ class ScienceMetadataRetrieveUpdate(APIView):
     allowed_methods = ('GET', 'PUT')
 
     def get(self, request, pk):
-        view_utils.authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.VIEW_METADATA)
+        authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.VIEW_METADATA)
 
         scimeta_url = hydroshare.utils.current_site_url() + AbstractResource.scimeta_url(pk)
         return redirect(scimeta_url)
 
     def put(self, request, pk):
         # Update science metadata based on resourcemetadata.xml uploaded
-        resource, authorized, user = view_utils.authorize(
+        resource, authorized, user = authorize(
             request, pk,
             needed_permission=ACTION_TO_AUTHORIZE.EDIT_RESOURCE,
             raises_exception=False)
@@ -687,7 +686,7 @@ class ResourceFileCRUD(APIView):
         return request
 
     def get(self, request, pk, filename):
-        view_utils.authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.VIEW_RESOURCE)
+        authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.VIEW_RESOURCE)
         try:
             f = hydroshare.get_resource_file(pk, filename)
         except ObjectDoesNotExist:
@@ -705,8 +704,7 @@ class ResourceFileCRUD(APIView):
         :param pk: Primary key of the resource (i.e. resource short ID)
         :return:
         """
-        resource, _, _ = view_utils.authorize(request, pk,
-                                              needed_permission=ACTION_TO_AUTHORIZE.EDIT_RESOURCE)
+        resource, _, _ = authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.EDIT_RESOURCE)
         resource_files = request.FILES.values()
         if len(resource_files) == 0:
             error_msg = {'file': 'No file was found to add to the resource.'}
@@ -747,8 +745,7 @@ class ResourceFileCRUD(APIView):
         return Response(data=response_data, status=status.HTTP_201_CREATED)
 
     def delete(self, request, pk, filename):
-        resource, _, user = view_utils.authorize(
-            request, pk, needed_permission=ACTION_TO_AUTHORIZE.EDIT_RESOURCE)
+        resource, _, user = authorize(request, pk, needed_permission=ACTION_TO_AUTHORIZE.EDIT_RESOURCE)
         try:
             hydroshare.delete_resource_file(pk, filename, user)
         except ObjectDoesNotExist as ex:    # matching file not found
@@ -817,8 +814,8 @@ class ResourceFileList(ResourceFileToListItemMixin, generics.ListAPIView):
         return self.list(request)
 
     def get_queryset(self):
-        resource, _, _ = view_utils.authorize(self.request, self.kwargs['pk'],
-                                              needed_permission=ACTION_TO_AUTHORIZE.VIEW_RESOURCE)
+        resource, _, _ = authorize(self.request, self.kwargs['pk'],
+                                   needed_permission=ACTION_TO_AUTHORIZE.VIEW_RESOURCE)
         resource_file_info_list = []
         for f in resource.files.all():
             resource_file_info_list.append(self.resourceFileToListItem(f))
